Extract <SampleUserGroup /> to separate module
